fix build

+++ TMessagesProj/src/main/java/org/telegram/messenger/fakepasscode/RemoveAsReadMessage.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 
 public class RemoveAsReadMessage {
     private int id;
-    private int topicId;
+    private long topicId;
     private long randomId = -1; // from encrypted dialogs
     private long readTime = -1;
     private int sendTime = -1;
@@ -13,7 +13,7 @@ public class RemoveAsReadMessage {
     public RemoveAsReadMessage() {
     }
 
-    public RemoveAsReadMessage(int id, int topicId, long randomId, int sendTime, int scheduledTimeMs) {
+    public RemoveAsReadMessage(int id, long topicId, long randomId, int sendTime, int scheduledTimeMs) {
         this.id = id;
         this.topicId = topicId;
         this.randomId = randomId;
@@ -29,7 +29,7 @@ public class RemoveAsReadMessage {
         this.id = id;
     }
 
-    public int getTopicId() {
+    public long getTopicId() {
         return topicId;
     }
